encoder: fix missing segment after redundant point
When encoding a point that is identical to a previous point,
that point is currently entirely skipped, as it is considered redundant.

However, this means that if in a multilinestring,
the first point of a line is identical to the last point of the previous line,
the second point of that line will be treated as the first.
Thus the segment from the end of the previous line to the second point
of the current line is not encoded.

This commit adds a flag to detect this case (in both multilinestring and multipolygons),
still skip the redundant MoveTo command,
but convert the moveTo command to the second point to a lineTo command.

Also add a test for this behavior as well as other forms of
redundant points.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>